fix: refactor deletePost controller

+++ controllers/posts.js
@@ -82,18 +82,23 @@ exports.putEditPost = async (req, res, next) => {
   }
 };
 
-exports.deletePost = (req, res, next) => {
+exports.deletePost = async (req, res, next) => {
   const postId = req.params.id;
-  Post.findByIdAndDelete(postId)
-    .then(() => {
-      console.log(postId);
-      console.log('Destroyed post');
+  const post = await Post.findById(postId);
+  if (post.userId.toString() !== req.user._id.toString()) {
+    return res.status(403).json({ message: "You are not the owner of this post" });
+  }
+  try {
+    const deletedPost = await Post.findByIdAndDelete(postId);
+    if (deletedPost) {
       res.status(200).json({ message: "Post deleted" });
-    })
-    .catch(err => {
-      console.error(err);
-      return res.status(500).json({ message: err.message });
-    })
+      console.log('Destroyed post');
+    } else {
+      res.status(404).json({ message: "Post not found" });
+    }
+  } catch (err) {
+    console.error(err);
+    return res.status(500).json({ message: 'Error deleting post' });
+  }
 
-  // return res.status(404).json({ message: "Post not found" });
 }
